requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-yasg==1.21.8
fastapi==0.115.8
greenlet==3.1.1
idna==3.10
inflection==0.5.1
Markdown==3.7
oauthlib==3.2.2
packaging==24.2
phonenumbers==8.13.53
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.37
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
